+++ src/Column/SequenceColumnSchemaTrait.php
@@ -9,8 +9,11 @@ trait SequenceColumnSchemaTrait
     /**
      * @var string|null Name of an associated sequence if column is auto incremental.
      */
-    private string|null $sequenceName = null;
+    protected string|null $sequenceName = null;
 
+    /**
+     * @psalm-mutation-free
+     */
     public function getSequenceName(): string|null
     {
         return $this->sequenceName;