+++ CHANGELOG.md
@@ -11,14 +11,16 @@
 - Enh #353: Update `bit` type according to main PR yiisoft/db#860 (@Tigrov) 
 - Enh #354: Refactor PHP type of `ColumnSchemaInterface` instances (@Tigrov)
 - Enh #356: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
-- New #355: Implement `ColumnFactory` class (@Tigrov)
+- New #355, #368, #370: Implement `ColumnFactory` class (@Tigrov)
 - Enh #359: Separate column type constants (@Tigrov)
 - Enh #359: Remove `Schema::TYPE_ARRAY` and `Schema::TYPE_STRUCTURED` constants (@Tigrov)
 - New #360: Realize `ColumnBuilder` class (@Tigrov)
 - Enh #362: Update according changes in `ColumnSchemaInterface` (@Tigrov)
-- New #364: Add `ColumnDefinitionBuilder` class (@Tigrov)
+- New #364, #372: Add `ColumnDefinitionBuilder` class (@Tigrov)
 - Enh #365: Refactor `Dsn` class (@Tigrov)
 - Enh #366: Use constructor to create columns and initialize properties (@Tigrov)
+- Enh #370: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
+- New #373: Override `QueryBuilder::prepareBinary()` method (@Tigrov)
 
 ## 1.3.0 March 21, 2024